Get a league-play tibble

Gets league-play data for each game, from engsoccerdata, returning a tibble in a standardised format.

uss_get_matches(country = uss_countries(), ...)

Arguments

country

character scalar, specifies the league. uss_countries() returns choices available from engsoccerdata

...

<data-masking> Expressions that return a logical value, and are defined in terms of the variables in .data. If multiple expressions are included, they are combined with the & operator. Only rows for which all conditions evaluate to TRUE are kept.

Value

tibble with columns country, date, season, tier, home, visitor, goals_home, goals_visitor.

Details

uss_countries() returns the available choices; "england" is the default.

This function relies on an internal function, uss_make_matches(), to parse the source data.

Examples

uss_get_matches("england")
#> # A tibble: 192,004 × 8
#>    country tier  season date       home            visitor       goals…¹ goals…²
#>    <chr>   <fct>  <int> <date>     <chr>           <chr>           <int>   <int>
#>  1 England 1       1888 1888-12-15 Accrington F.C. Aston Villa         1       1
#>  2 England 1       1888 1889-01-19 Accrington F.C. Blackburn Ro…       0       2
#>  3 England 1       1888 1889-03-23 Accrington F.C. Bolton Wande…       2       3
#>  4 England 1       1888 1888-12-01 Accrington F.C. Burnley             5       1
#>  5 England 1       1888 1888-10-13 Accrington F.C. Derby County        6       2
#>  6 England 1       1888 1888-12-29 Accrington F.C. Everton             3       1
#>  7 England 1       1888 1889-01-26 Accrington F.C. Notts County        1       2
#>  8 England 1       1888 1888-10-20 Accrington F.C. Preston Nort…       0       0
#>  9 England 1       1888 1889-04-20 Accrington F.C. Stoke City          2       0
#> 10 England 1       1888 1888-11-24 Accrington F.C. West Bromwic…       2       1
#> # … with 191,994 more rows, and abbreviated variable names ¹​goals_home,
#> #   ²​goals_visitor
#> # ℹ Use `print(n = ...)` to see more rows
uss_get_matches("italy", goals_home ==2)
#> # A tibble: 6,035 × 8
#>    country tier  season date       home                visitor   goals…¹ goals…²
#>    <chr>   <fct>  <int> <date>     <chr>               <chr>       <int>   <int>
#>  1 Italy   1       1934 1934-09-30 Sampierdarenese     Bologna …       2       1
#>  2 Italy   1       1934 1934-10-07 Unione Triestina    Lazio Ro…       2       1
#>  3 Italy   1       1934 1934-10-07 US Alessandria 1912 Sampierd…       2       2
#>  4 Italy   1       1934 1934-10-07 Juventus            SSC Napo…       2       1
#>  5 Italy   1       1934 1934-10-07 US Palermo          Pro Verc…       2       0
#>  6 Italy   1       1934 1934-10-14 Lazio Roma          Bologna …       2       1
#>  7 Italy   1       1934 1934-10-14 Torino FC           US Aless…       2       1
#>  8 Italy   1       1934 1934-10-14 Brescia Calcio      AC Milan        2       0
#>  9 Italy   1       1934 1934-10-14 US Palermo          US Livor…       2       2
#> 10 Italy   1       1934 1934-10-21 AS Roma             Sampierd…       2       0
#> # … with 6,025 more rows, and abbreviated variable names ¹​goals_home,
#> #   ²​goals_visitor
#> # ℹ Use `print(n = ...)` to see more rows