Create a standard dataframe for league data

Create a standard dataframe for league data

Usage

uss_make_matches(data_engsoc, country_name)

Arguments

data_engsoc

Data produced by engsoccerdata

country_name

Name of country that will be added to output to identify country used

Value

a tibble with columns country, date, season, tier, home, visitor, goals_home, goals_visitor.

Examples

uss_make_matches(engsoccerdata::england, country_name = "England")
#> # A tibble: 192,004 × 8
#>    country tier  season date       home            visitor       goals…¹ goals…²
#>    <chr>   <fct>  <int> <date>     <chr>           <chr>           <int>   <int>
#>  1 England 1       1888 1888-12-15 Accrington F.C. Aston Villa         1       1
#>  2 England 1       1888 1889-01-19 Accrington F.C. Blackburn Ro…       0       2
#>  3 England 1       1888 1889-03-23 Accrington F.C. Bolton Wande…       2       3
#>  4 England 1       1888 1888-12-01 Accrington F.C. Burnley             5       1
#>  5 England 1       1888 1888-10-13 Accrington F.C. Derby County        6       2
#>  6 England 1       1888 1888-12-29 Accrington F.C. Everton             3       1
#>  7 England 1       1888 1889-01-26 Accrington F.C. Notts County        1       2
#>  8 England 1       1888 1888-10-20 Accrington F.C. Preston Nort…       0       0
#>  9 England 1       1888 1889-04-20 Accrington F.C. Stoke City          2       0
#> 10 England 1       1888 1888-11-24 Accrington F.C. West Bromwic…       2       1
#> # … with 191,994 more rows, and abbreviated variable names ¹​goals_home,
#> #   ²​goals_visitor
#> # ℹ Use `print(n = ...)` to see more rows