Make a standard league-play tibble

Given a league-play data frame from engsoccer, returns a tibble with standardised colomn-names and types, e.g. date is a Date.

There are two versions of this function:

uss_make_matches()

working version

uss_make_matches_mem()

memoised (caching) version

uss_make_matches_mem() calls uss_make_matches(); they will resturn identical results.

Usage

uss_make_matches(data_engsoc, country)

uss_make_matches_mem(data_engsoc, country)

Arguments

data_engsoc

data.frame obtained from engsoccerdata.

country

character scalar, specifies the league. uss_countries() returns choices available from engsoccerdata

Value

tibble with columns country, date, season, tier, home, visitor, goals_home, goals_visitor.

Details

A memoised version of this function is provided because it takes a little time to parse the dates for nearly 200,000 league-play matches for England.

When the package is loaded, the wrapper function, uss_make_matches_mem(), is "replaced" by the memoised version: see zzz.R.

A wrapper function is used for a couple reasons:

to maintain access to the un-memoised function

so that R CMD CHECK can see that the tibble package is actually used; it throws a note if the original function, uss_make_matches_mem(), is overwritten

Examples

uss_make_matches(engsoccerdata::italy, "Italy")
#> # A tibble: 25,404 × 8
#>    country tier  season date       home         visitor goals_home goals_visitor
#>    <chr>   <fct>  <int> <date>     <chr>        <chr>        <int>         <int>
#>  1 Italy   1       1934 1934-09-30 Lazio Roma   US Liv…          6             1
#>  2 Italy   1       1934 1934-09-30 Torino FC    Unione…          3             1
#>  3 Italy   1       1934 1934-09-30 Sampierdare… Bologn…          2             1
#>  4 Italy   1       1934 1934-09-30 SSC Napoli   US Ale…          0             1
#>  5 Italy   1       1934 1934-09-30 ACF Fiorent… AS Roma          4             1
#>  6 Italy   1       1934 1934-09-30 Brescia Cal… Juvent…          0             2
#>  7 Italy   1       1934 1934-09-30 Inter        US Pal…          3             0
#>  8 Italy   1       1934 1934-09-30 Pro Vercelli AC Mil…          1             2
#>  9 Italy   1       1934 1934-10-07 AC Milan     ACF Fi…          1             1
#> 10 Italy   1       1934 1934-10-07 US Livorno   Inter            1             1
#> # … with 25,394 more rows
uss_make_matches_mem(engsoccerdata::italy, "Italy")
#> # A tibble: 25,404 × 8
#>    country tier  season date       home         visitor goals_home goals_visitor
#>    <chr>   <fct>  <int> <date>     <chr>        <chr>        <int>         <int>
#>  1 Italy   1       1934 1934-09-30 Lazio Roma   US Liv…          6             1
#>  2 Italy   1       1934 1934-09-30 Torino FC    Unione…          3             1
#>  3 Italy   1       1934 1934-09-30 Sampierdare… Bologn…          2             1
#>  4 Italy   1       1934 1934-09-30 SSC Napoli   US Ale…          0             1
#>  5 Italy   1       1934 1934-09-30 ACF Fiorent… AS Roma          4             1
#>  6 Italy   1       1934 1934-09-30 Brescia Cal… Juvent…          0             2
#>  7 Italy   1       1934 1934-09-30 Inter        US Pal…          3             0
#>  8 Italy   1       1934 1934-09-30 Pro Vercelli AC Mil…          1             2
#>  9 Italy   1       1934 1934-10-07 AC Milan     ACF Fi…          1             1
#> 10 Italy   1       1934 1934-10-07 US Livorno   Inter            1             1
#> # … with 25,394 more rows