Make a standard league-play tibble

Given a league-play data frame from engsoccer, returns a tibble with standardised colomn-names and types, e.g. date is a Date.

uss_make_matches(data_engsoc, country)

Arguments

data_engsoc

data.frame obtained from engsoccerdata.

country

character scalar, specifies the league. uss_countries() returns choices available from engsoccerdata

Value

tibble with columns country, date, season, tier, home, visitor, goals_home, goals_visitor.

Examples

uss_make_matches(engsoccerdata::italy, "Italy")
#> # A tibble: 25,404 × 8
#>    country tier  season date       home            visitor       goals…¹ goals…²
#>    <chr>   <fct>  <int> <date>     <chr>           <chr>           <int>   <int>
#>  1 Italy   1       1934 1934-09-30 Lazio Roma      US Livorno          6       1
#>  2 Italy   1       1934 1934-09-30 Torino FC       Unione Tries…       3       1
#>  3 Italy   1       1934 1934-09-30 Sampierdarenese Bologna FC          2       1
#>  4 Italy   1       1934 1934-09-30 SSC Napoli      US Alessandr…       0       1
#>  5 Italy   1       1934 1934-09-30 ACF Fiorentina  AS Roma             4       1
#>  6 Italy   1       1934 1934-09-30 Brescia Calcio  Juventus            0       2
#>  7 Italy   1       1934 1934-09-30 Inter           US Palermo          3       0
#>  8 Italy   1       1934 1934-09-30 Pro Vercelli    AC Milan            1       2
#>  9 Italy   1       1934 1934-10-07 AC Milan        ACF Fiorenti…       1       1
#> 10 Italy   1       1934 1934-10-07 US Livorno      Inter               1       1
#> # … with 25,394 more rows, and abbreviated variable names ¹​goals_home,
#> #   ²​goals_visitor
#> # ℹ Use `print(n = ...)` to see more rows